refactor: rename helpers/dummy to helpers/api
